Remove old error emitter
This completes the transition to annotate-snippets
