+++ compiler/rustc_parse/src/parser/tests.rs
@@ -12,7 +12,7 @@ use rustc_ast::{self as ast, PatKind, visit};
 use rustc_ast_pretty::pprust::item_to_string;
 use rustc_data_structures::assert_matches;
 use rustc_errors::annotate_snippet_emitter_writer::AnnotateSnippetEmitter;
-use rustc_errors::emitter::{HumanEmitter, OutputTheme};
+use rustc_errors::emitter::OutputTheme;
 use rustc_errors::translation::Translator;
 use rustc_errors::{AutoStream, DiagCtxt, MultiSpan, PResult};
 use rustc_session::parse::ParseSess;
@@ -49,20 +49,12 @@ fn create_test_handler(theme: OutputTheme) -> (DiagCtxt, Arc<SourceMap>, Arc<Mut
     let translator = Translator::with_fallback_bundle(vec![crate::DEFAULT_LOCALE_RESOURCE], false);
     let shared: Box<dyn Write + Send> = Box::new(Shared { data: output.clone() });
     let auto_stream = AutoStream::never(shared);
-    let dcx = DiagCtxt::new(match theme {
-        OutputTheme::Ascii => Box::new(
-            HumanEmitter::new(auto_stream, translator)
-                .sm(Some(source_map.clone()))
-                .diagnostic_width(Some(140))
-                .theme(theme),
-        ),
-        OutputTheme::Unicode => Box::new(
-            AnnotateSnippetEmitter::new(auto_stream, translator)
-                .sm(Some(source_map.clone()))
-                .diagnostic_width(Some(140))
-                .theme(theme),
-        ),
-    });
+    let dcx = DiagCtxt::new(Box::new(
+        AnnotateSnippetEmitter::new(auto_stream, translator)
+            .sm(Some(source_map.clone()))
+            .diagnostic_width(Some(140))
+            .theme(theme),
+    ));
     (dcx, source_map, output)
 }
